[IMP] loca mexico: update down payment documentation v18.0
"This is no longer configured in v17.4 or v18.0, as a default label is already generated with the necessary values for the CFDI."

First, navigate to the :menuselection:`Sales` app to create a product `Anticipo` and configure it.
The :guilabel:`Product Type` must be :guilabel:`Service`, and use the :guilabel:`UNSPSC Category` must be: `84111506 Servicios de facturación`.

Then, go to :menuselection:`Sales --> Settings --> Invoicing --> Down Payments`, and add the *Anticipo* product as the default.

"The correct order is to first apply the down payment to the new invoice and then register the payments so that the outstanding balance is properly documented."

With this, all electronic documents are linked to each other. The final step is to fully pay the new invoice. At the bottom of the new invoice, you can find the credit note in the  :guilabel:`Outstanding credits` - add it as payment. Finally, register the remaining amount with the :guilabel:`Register Payment` wizard.